Closures were giving me a headache so I removed them.
The closures that I implemented previously were
untested and frankly broken from their inception.
I theorized that I could use type information to
know the type of each closure enviroment as a tuple,
but this doesn't allow for closures with different
environment types to be passed in to a function such
as map.

I still think closures are useful, but given I
have yet to reconcile them with linear/affine types
I am going to wait to reintroduce them until I have
a better feel on the memory model of the language
(or until I give up and just write a garbage collector)