closures partially reimplemented
added closures represented as (ptr, ptr),
for now all environments are on the heap

referencing a function name is broken as
I am not creating a closure when I see that identifier.

likely I'm going to modify call to check if the env is null,
if it is we just call with args, otherwise we call with env.

alternatively I could create a generic shim function, but I don't
like having so many thunk methods created for every type/number of args.