Scraping the rust off the last commits...
by actually testing the code. I mean I'm
not a genius, but I at least thought I
would be smart enought to ACTUALLY TEST
MY OWN CHANGES!!! But as it turns out
everything was broken in a lot of ways,
not thanks to my habit of changing
everything at once. Here's hoping the
rewrite goes better than this *shudder*